# src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description: Single Sign On service for V.A.P.E
  title: V.A.P.E S.S.O
  version: 0.0.1 snapshot
host: localhost:8821
basePath: /sso/v1
schemes:
  - http
  - https
paths:
  '/token/client':
    get:
      tags:
        - Check if session is active
      summary: check if a session is alive
      operationId: getSession
      produces:
        - application/json

      parameters:
        - in: query
          name: sessionId
          required: true
          type: string
          description: The id of session to check
        - in: query
          name: token
          required: true
          type: string
          description: The token of session to check

      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/TokenState'
          headers:
            Access-Control-Allow-Origin:
              type: string
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/BadRequestException'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/Exception'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Exception'

  '/token/client/{clientId}':
    post:
      tags:
        - Activate new session
      operationId: PostNewSession
      summary: Activate new session
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: clientId
          required: true
          type: string
          description: The id of session to check
        - in: body
          name: Credential
          required: true
          schema:
            $ref: '#/definitions/SessionRequest'
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/Token'
          headers:
            Access-Control-Allow-Origin:
              type: string
              description: allowed url
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/BadRequestException'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/Exception'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Exception'

definitions:
  Token:
    type: object
    properties:
      sessionId:
        type: string
      secrete:
        type: string
  SessionRequest:
    type: object
    properties:
      client-id:
        type: string
      client-secret:
        type: string
      user-agent:
        type: string
  TokenState:
    type: object
    properties:
      status:
        $ref: '#/definitions/TokenStatus'
  TokenStatus:
    type: string
    enum:
      - EXPIRED
      - INVALID
      - ACTIVE
  BadRequestException:
    type: object
    properties:
      message:
        type: integer
      errorCode:
        type: string
        readOnly: false
    required:
      - message
      - errorCode
  Exception:
    type: object
    properties:
      message:
        type: integer
      errorCode:
        type: string
        readOnly: false
    required:
      - message
      - errorCode